@@ -5,6 +5,7 @@ namespace App;
 use Eloquent;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Collection;
+use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\Relations\HasMany;
 use Illuminate\Foundation\Auth\User as Authenticable;
@@ -64,7 +65,8 @@ use Spatie\Permission\Traits\HasRoles;
  */
 class User extends Authenticable
 {
-    use HasRoles, HasApiTokens, Notifiable;
+    use HasRoles, Notifiable, HasFactory;
+
 
     /**
      * The attributes that are mass assignable.
