@@ -1,13 +1,28 @@
 <?php
 
-/* @var $factory Factory */
+namespace Database\Factories;
 
 use App\BundleArticle;
-use Faker\Generator as Faker;
-use Illuminate\Database\Eloquent\Factory;
+use Illuminate\Database\Eloquent\Factories\Factory;
 
-$factory->define(BundleArticle::class, function (Faker $faker) {
-    return [
-        //
-    ];
-});
+class BundleArticleFactory extends Factory
+{
+    /**
+     * The name of the factory's corresponding model.
+     *
+     * @var string
+     */
+    protected $model = BundleArticle::class;
+
+    /**
+     * Define the model's default state.
+     *
+     * @return array
+     */
+    public function definition(): array
+    {
+        return [
+            //
+        ];
+    }
+}
