@@ -1,14 +1,30 @@
 <?php
 
-/* @var $factory Factory */
+namespace Database\Factories;
 
 use App\Bundle;
-use Faker\Generator as Faker;
-use Illuminate\Database\Eloquent\Factory;
+use App\Supplier;
+use Illuminate\Database\Eloquent\Factories\Factory;
 
-$factory->define(Bundle::class, function (Faker $faker) {
-    return [
-        'name' => $faker->colorName,
-        'quantity' => $faker->numberBetween(0, 100),
-    ];
-});
+class BundleFactory extends Factory
+{
+    /**
+     * The name of the factory's corresponding model.
+     *
+     * @var string
+     */
+    protected $model = Bundle::class;
+
+    /**
+     * Define the model's default state.
+     *
+     * @return array
+     */
+    public function definition(): array
+    {
+        return [
+            'name' => $this->faker->colorName,
+            'quantity' => $this->faker->numberBetween(0, 100),
+        ];
+    }
+}
