@@ -1,13 +1,28 @@
 <?php
 
-/* @var $factory Factory */
+namespace Database\Factories;
 
 use App\Price;
-use Faker\Generator as Faker;
-use Illuminate\Database\Eloquent\Factory;
+use Illuminate\Database\Eloquent\Factories\Factory;
 
-$factory->define(Price::class, function (Faker $faker) {
-    return [
-        'value' => $faker->randomFloat(2, 0, 15.99)
-    ];
-});
+class PriceFactory extends Factory
+{
+    /**
+     * The name of the factory's corresponding model.
+     *
+     * @var string
+     */
+    protected $model = Price::class;
+
+    /**
+     * Define the model's default state.
+     *
+     * @return array
+     */
+    public function definition(): array
+    {
+        return [
+            'value' => $this->faker->randomFloat(2, 0, 15.99)
+        ];
+    }
+}
